--- a/lib/reactotron-mst/package.json
+++ b/lib/reactotron-mst/package.json
@@ -82,7 +82,6 @@
     "rollup-plugin-filesize": "^6.0.1",
     "rollup-plugin-node-resolve": "^4.0.0",
     "rollup-plugin-resolve": "^0.0.1-predev.1",
-    "testdouble": "^3.20.0",
     "ts-jest": "^29.1.1",
     "typescript": "^4.9.5"
   },