--- a/lib/reactotron-mst/test/fixtures/create-mst-plugin.ts
+++ b/lib/reactotron-mst/test/fixtures/create-mst-plugin.ts
@@ -1,4 +1,5 @@
-import { mst, MstPluginOptions } from "../../src/reactotron-mst"
+import type { Reactotron } from "reactotron-core-client"
+import { mst, type MstPluginOptions } from "../../src/reactotron-mst"
 import { createMockReactotron } from "../mocks/create-mock-reactotron"
 
 /**
@@ -6,7 +7,7 @@ import { createMockReactotron } from "../mocks/create-mock-reactotron"
  */
 export function createMstPlugin(pluginOptions: MstPluginOptions = {}) {
   const reactotron = createMockReactotron()
-  const plugin = mst(pluginOptions)(reactotron)
+  const plugin = mst(pluginOptions)(reactotron as unknown as Reactotron)
   const track = plugin.features.trackMstNode
 
   return { reactotron, plugin, track }