This is in preparation for unlocking bun test in the future.

`testdouble` creates a cryptic package error when running in `bun test` on this workspace.

This PR refactors to use jest.fn and other mocking features instead of `testdouble`.

| Before | After |
|--------|--------|
| <img width="868" alt="Screenshot 2023-12-08 at 2 32 59 PM" src="https://github.com/infinitered/reactotron/assets/37849890/d9905795-8194-48e2-ab1e-27742be914f6"> | <img width="496" alt="Screenshot 2023-12-08 at 2 33 47 PM" src="https://github.com/infinitered/reactotron/assets/37849890/bf43e21b-f419-4df8-82d2-75806d3dd651"> |

Bun is 200x times faster than Jest 🤯 

2s isn't bad for Jest for these tests but I think adding them up across the repo will help speed up the CI a lot.