@joshuayoes 2023-12-08
In the future, migrating this suite to bun will be a 1 liner change
<img width="542" alt="Screenshot 2023-12-08 at 2 47 21 PM" src="https://github.com/infinitered/reactotron/assets/37849890/4dcd6aa1-151b-42df-9b7d-366490deb1f0">